Basic styling to sell the idea better
